// types.ts
export interface BillingConfig {
  isPaywallActive: boolean;
  monthlyFee: number;
  durationDays: number;
  contactZalo?: string; // Mới thêm vào
  bankInfo: {
    bankName: string;
    accountNumber: string;
    accountName: string;
  };
}

export interface OfficeLocation {
  id: string;
  name: string;
  lat: number;
  lng: number;
  radius: number;
  qrCodeString: string;
}

export interface Company {
  id: string;
  name: string;
  expiredAt: string; // ISO string
  createdAt: string;
  exportCompanyName?: string;
  taxCode?: string;
  phone?: string;
  address?: string;
}